requirements.txt
PyGithub>=2.1.1
langchain>=0.1.16
langchain-openai>=0.1.3
langchain-google-genai>=1.0.3
python-dotenv>=1.0.1
streamlit>=1.32.0